requirements.txt
pyyaml
paramiko
progressbar2
